+++ src/pychoreo_examples/extrusion/visualization.py
@@ -2,7 +2,7 @@ import time
 import numpy as np
 import colorsys
 from pybullet_planning import LockRenderer, set_camera_pose, add_line, add_text, wait_for_user, connect, load_pybullet, \
-    reset_simulation, disconnect, wait_for_duration
+    reset_simulation, disconnect, wait_for_duration, HideOutput
 from pybullet_planning import joints_from_names, set_joint_positions
 from pybullet_planning import get_link_pose, link_from_name
 from pybullet_planning import unit_pose, multiply, tform_point, point_from_pose
@@ -49,6 +49,7 @@ def draw_extrusion_sequence(node_points, element_bodies, element_sequence, seq_p
     handles = []
 
     for seq_id, element in enumerate(element_sequence):
+        print('visualizing EE directions for seq #{}'.format(seq_id))
         n1, n2 = element
         p1, p2 = (node_points[n1], node_points[n2])
         e_mid = (np.array(p1) + np.array(p2)) / 2
@@ -78,13 +79,14 @@ def draw_extrusion_sequence(node_points, element_bodies, element_sequence, seq_p
 ##################################################
 
 def display_trajectories(robot_urdf, ik_joint_names, ee_link_name, node_points, ground_nodes, trajectories,
-                         workspace_urdf=None, animate=True, time_step=0.02):
+                         workspace_urdf=None, animate=True, cart_time_step=0.02, tr_time_step=0.05):
     if trajectories is None:
         return
     connect(use_gui=True)
     set_extrusion_camera(node_points)
-    robot = load_pybullet(robot_urdf, fixed_base=True)
-    if workspace_urdf: workspace = load_pybullet(workspace_urdf, fixed_base=True)
+    with HideOutput():
+        robot = load_pybullet(robot_urdf, fixed_base=True)
+        if workspace_urdf: workspace = load_pybullet(workspace_urdf, fixed_base=True)
     ik_joints = joints_from_names(robot, ik_joint_names)
 
     if not animate:
@@ -117,12 +119,19 @@ def display_trajectories(robot_urdf, ik_joint_names, ee_link_name, node_points, 
                     color = (0, 0, 1) if is_ground(trajectory.element, ground_nodes) else (1, 0, 0)
                     handles.append(add_line(last_point, current_point, color=color))
                 last_point = current_point
-            if time_step is None:
-                wait_for_user()
+                if cart_time_step is None:
+                    wait_for_user()
+                else:
+                    # ! this seems to have a bug on windows
+                    # wait_for_duration(time_step)
+                    time.sleep(cart_time_step)
             else:
-                # ! this seems to have a bug on windows
-                # wait_for_duration(time_step)
-                time.sleep(time_step)
+                if tr_time_step is None:
+                    wait_for_user()
+                else:
+                    # ! this seems to have a bug on windows
+                    # wait_for_duration(time_step)
+                    time.sleep(tr_time_step)
 
         # * sanity check on connectness
         if isinstance(trajectory, PrintTrajectory):