+++ tests/conftest.py
@@ -1,8 +1,7 @@
 import pytest
 
-
 def pytest_addoption(parser):
-    parser.addoption("--viewer", action="store_true", help='Enables the pybullet viewer')
+    parser.addoption('--viewer', action='store_true', help='Enables the pybullet viewer')
 
 @pytest.fixture
 def viewer(request):